LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/response-null-body.any.worker-expected.txt


FAIL Response.body is null for responses with status=204 (method=GET) assert_equals: the body should be null expected null but got object "[object ReadableStream]"
FAIL Response.body is null for responses with status=204 (method=POST) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with status=204 (method=OPTIONS) assert_equals: the body should be null expected null but got object "[object ReadableStream]"
FAIL Response.body is null for responses with status=205 (method=GET) assert_equals: the body should be null expected null but got object "[object ReadableStream]"
FAIL Response.body is null for responses with status=205 (method=POST) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with status=205 (method=OPTIONS) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with status=304 (method=GET) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with status=304 (method=POST) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with status=304 (method=OPTIONS) promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Response.body is null for responses with method=HEAD promise_test: Unhandled rejection with value: object "TypeError: Load failed"
PASS Null body status with subresource integrity should abort
